src/main/java/io/github/hrashk/news/api/authors/AuthorService.java
package io.github.hrashk.news.api.authors;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class AuthorService {
    private final AuthorRepository repository;

    public List<Author> findAll() {
        return repository.findAll();
    }

    public Author save(Author author) {
        return repository.save(author);
    }

    public Author findById(Long id) {
        return repository.findById(id).orElseThrow();
    }
}
